requirements.txt
aioredis==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
autoflake==2.3.1
bcrypt==4.0.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
fastapi-versioning==0.10.0
flake8==7.2.0
flower==2.0.1
fonttools==4.56.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
meson==1.6.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg-binary==3.2.3
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.3.2
PyJWT==2.10.1
pyparsing==3.2.1
pyright==1.1.399
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.19
pytz==2025.1
redis==5.2.1
rsa==4.9
sentry-sdk==2.25.1
setuptools==75.6.0
setuptools-metadata==0.1.5
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
subprocess.run==0.0.8
Tcl==0.2
time-machine==2.16.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
WTForms==3.1.2
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
